pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "discord_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.10.9"
aiosignal = "1.3.1"
async-timeout = "4.0.3"
attrs = "24.2.0"
beautifulsoup4 = "4.12.3"
brotli = "1.1.0"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.3.2"
discord-py = "2.4.0"
frozenlist = "1.4.1"
idna = "3.10"
multidict = "6.1.0"
mutagen = "1.47.0"
pycparser = "2.22"
pycryptodomex = "3.21.0"
pynacl = "1.5.0"
python-dotenv = "1.0.1"
requests = "2.32.3"
soupsieve = "2.6"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
websockets = "13.1"
yarl = "1.13.1"
yt-dlp = "2024.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
